+++ src/main/java/net/silentchaos512/loginar/data/ModBlockStateProvider.java
@@ -3,6 +3,7 @@ package net.silentchaos512.loginar.data;
 import net.minecraft.data.DataGenerator;
 import net.minecraftforge.client.model.generators.BlockStateProvider;
 import net.minecraftforge.common.data.ExistingFileHelper;
+import net.silentchaos512.lib.util.NameUtils;
 import net.silentchaos512.loginar.LoginarMod;
 import net.silentchaos512.loginar.block.urn.LoginarUrnBlock;
 import net.silentchaos512.loginar.setup.LsBlocks;
@@ -16,7 +17,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
     protected void registerStatesAndModels() {
         for (LoginarUrnBlock block : LsBlocks.getUrns()) {
             simpleBlock(block, models()
-                    .getExistingFile(modLoc("block/loginar_urn"))
+                    .getExistingFile(modLoc("block/" + NameUtils.fromBlock(block).getPath()))
             );
         }
     }