+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Each loginar urn size now has a unique model. The large urn retains the original model.
+
 ## [0.2.1] - 2022-12-22
 ### Added
 - Item swapper upgrade. By pressing a key (default X), opens a menu that allows the player to swap their held item with a selected item from within the urn. Does not work when holding an urn.